+++ app/Http/Controllers/Candidat/CandidatureController.php
@@ -30,7 +30,7 @@ class CandidatureController extends Controller
 
     public function detail(Candidature $candidature)
     {
-        abort_if($candidature->candidat_id !== Auth::id(), 403);
+        abort_if((int) $candidature->candidat_id !== (int) Auth::id(), 403);
         $candidature->load(['offre.recruteur', 'cv', 'documents.type']);
         return view('candidat.candidature-detail', compact('candidature'));
     }
